refactor confusing name to ComparisonContainer
